requirements.txt
auto-py-to-exe
bcrypt
beautifulsoup4
browser-cookie3
colorama
cryptography
customtkinter
deep-translator
discord
dnspython
exifread
GPUtil
keyboard
mouse
moviepy
opencv-python
pdf2docx
phonenumbers
piexif
pillow
psutil
pyautogui
pycryptodome
pydub
pyinstaller
pyqt5
pyqtwebengine
pywin32
pyzipper
python-docx
rarfile
requests
screeninfo
selenium
setuptools
urllib3
whois
tqdm
numpy
pyperclip
pystyle
pygetwindow
pyglet
qrcode
matplotlib
pytesseract
scikit-image
discord.py
pypresence
pygame
rich
pytube
asyncio
discord-webhook==1.4.1